requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
flaml[automl]
shap
langchain
google-generativeai
python-dotenv
langchain-google-genai
